requirements/base.txt
Django==4.1
Jinja2==3.0.3
Pillow==9.0.0
djangorestframework==3.13.1
redis==4.3.1
celery==5.2.7
environs==9.3.0
psycopg2==2.9.3
django-timestampable==1.0.3
django-mptt==0.13.4
transliterate==1.10.2
pydantic==1.10.2
requests==2.28.1
